added a delete mapping

+++ jetty/src/main/java/com/embeddedcontainer/jetty/Repository/TodoRepository.java
@@ -4,6 +4,6 @@ import com.embeddedcontainer.jetty.Entity.TodoEntity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodoRepository extends JpaRepository<TodoEntity, Integer> {
+public interface TodoRepository extends JpaRepository<TodoEntity, Long> {
 
 }
